wyhash.h: add fallback definitions
This allows wyhash to work with compilers without builtins or with
different or none endianess detection macros.
Note that #warning is not recognised by all compilers, those that don't
and will instead error out when endianess could not be detected. If
endianess could be detected, it does not matter whether or not the
compiler recognises #warning.

The fallback for Hemming Weight computation is a well known to be
efficient implementation.